pipewire-object-mixin: use 64-bit type for `change_mask`
All core `pw_*_info` structures have a 64-bit change mask.
However, the `change_mask` in the session manager extension
info structures is only 32-bit. The differing sizes can cause
problems, especially on big endian architectures at the moment.

See #49

+++ lib/wp/private/pipewire-object-mixin.h
@@ -54,9 +54,9 @@ struct _WpPwObjectMixinPrivInterface
   gsize param_info_offset;
   gsize n_params_offset;
 
-  guint32 CHANGE_MASK_ALL;
-  guint32 CHANGE_MASK_PROPS;
-  guint32 CHANGE_MASK_PARAMS;
+  guint64 CHANGE_MASK_ALL;
+  guint64 CHANGE_MASK_PROPS;
+  guint64 CHANGE_MASK_PARAMS;
 
   gpointer (*update_info) (gpointer info, gconstpointer update);
   void (*free_info) (gpointer info);
